dpa_offload: Extend the IPSec support for SA modify
The requirements is needed in order to support IPSec High Availability

DPAA Offload IPSec API is now supporting sequence number &
extended sequence number updates.

This operation is done asynchronous the SEC is executing a RJD
which modifies the ESN as required. Notifications will be done
through a FD with user error status and in frame will be a code
that specifies the opperation and the SA id that suffered the
opperation.

Change-Id: I785dbbdb54b5df56c2a5ebd244bea7fdccfc0c92
Signed-off-by: Andrei Varvara <andrei.varvara@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/6038
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>