dpaa_eth: Add optimization option for jumbo frames
Termination traffic with large frames can perform better if only
linear skbs (and single buffer frames) are used. So add a Kconfig
option that allows the driver to allocate Rx buffers large enough
to accomodate the entire frame (of max 9.6K).

This option needs to be used with caution and only for jumbo frames
termination traffic, as the memory footprint can be a real problem
in case of small frames.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: I32d48cf0374622be5aeed7c83b6b7d544a3929ef
Reviewed-on: http://git.am.freescale.net:8181/6148
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>