dpaa_eth: macless control over mac device
This patch adds a new feature to macless DPAA Ethernet Driver. The macless
interface is now able to control a mac device through proxy DPAA Ethernet
Driver. To do that, it maintains a reference to a proxy interface.

The proxy interface offers a simple API to macless interface for start/stop and
for adding and removing mac unicast and multicast addresses.

Additionally, this patch will generalize some of mac API handlers. In previous
implementation, the mac handler and net_device structure belong to the same
network interface. With this patch, mac handler and net_device structure could
belong to different network interfaces. This is needed for the new macless
multicast capability, when the proxy interface that contains the mac device
reference should configure multicast addresses on behalf of a macless
interface.

This feature is optional and can be activated by setting "proxy" attribute in
the device tree node of the macless interfaces. Look in feature design document
for more details.

Change-Id: I92bf506324a9a834fdf27cbf989df817751b446d
Signed-off-by: Marian Rotariu <marian.rotariu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/6102
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Ruxandra Ioana Radulescu <ruxandra.radulescu@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>