KVM: PPC: e500: TLB emulation for IND entries
Handle indirect entries (IND) in TLB emulation code. Translation size of IND
entries differ from the size of referred Page Tables (Linux guests now use IND
of 2MB for 4KB PTs) and this require careful tweak of the existing logic.

TLB search emulation requires additional search in HW TLB0 (since these entries
are directly added by HTW) and found entries shoud be presented to the guest with
RPN changed from PFN to GFN. There might be more GFNs pointing to the same PFN so
the only way to get the corresponding GFN is to search it in guest's PTE. If IND
entry for the corresponding PT is not available just invalidate guest's ea and
report a tlbsx miss. This patch only implements the invalidation and let a TODO
note for searching HW TLB0.

Change-Id: I181afc1e8046b8858d4cc37398d7115c9cc85318
Signed-off-by: Mihai Caraman <mihai.caraman@freescale.com>

Conflicts:
	arch/powerpc/kvm/e500_mmu_host.c
Change-Id: I65e0bd112bd94116793858c4f92fb374fbc42d59
Reviewed-on: http://git.am.freescale.net:8181/6474
Tested-by: Richard Schmitt <B43082@freescale.com>
Reviewed-by: Richard Schmitt <B43082@freescale.com>