fmd: Fix problem in dynamic resource allocation algorithm
The patch that removed the forwarding optimized DPAA Ethernet driver
(and implicitly the Kconfig option CONFIG_FSL_DPAA_ETH_SG_SUPPORT and
all associated code) deleted the wrong branch of an #ifdef in the fmd
dynamic resource allocation algorithm.

As a result, the allocated FIFO size was not computed correctly,
causing resource depletion in case jumbo frames support is enabled.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: Idbd8805a5b475ab9b54005e6782e8f8361122540
Reviewed-on: http://git.am.freescale.net:8181/6544
Reviewed-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>