powerpc/p1010rdb-pb:remove phy interrupts from p1010rdb-pb
The function autoneg can in interrupt mode and polling mode.
On board p1010rdb-pb, there is an issue leading bad work in interrupt mode.
There are no phy-interrupts generated when cable is plugged in and out.
So remove phy interrupts to let it work in polling mode as a workaround.

Signed-off-by: Zhao Qiang <B45475@freescale.com>
Change-Id: I1ecd650b8aa116521d6d1c8a694013b5543cb304
Reviewed-on: http://git.am.freescale.net:8181/6892
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Claudiu Manoil <claudiu.manoil@freescale.com>
Reviewed-by: Tiefei Zang <roy.zang@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>