fsl_pci: Align PCICSRBAR to not include BAR0 attribute
BAR0 is a 32 bit register which maps to CCSR memory.
Lower bits of BARs are attributes associated with that BAR and
have following meaning:
 - bit-0: 0-IO space; 1-Memory Space
 - bit-1: if IO-space it is reserved;
   bit-2: if memory space then bit2-bit1 defines Locatable type
 - bit-3: if memory space then it is Prefetchable attribute

So when reading BAR0 (in fact all bars) for base address then we should
mask out these attribute bits. IO space mask is defined as
PCI_BASE_ADDRESS_IO_MASK and memory space mask is defined as
PCI_BASE_ADDRESS_MEM_MASK.

Now for BAR0, we always treat this as memory space (PCI_BASE_ADDRESS_SPACE_MEMORY)
because this seems to broken on some FSL SOCs(T4240-rev2 etc).
T4240-Rev2 reports attributes as
 - Bit0 == 1 : IO space; I think it is incorrect
 - Bit1 == 1 : Looks like incorrect again as it is IO space and bit1 is reserved
                and I think it should be 0.
 - bit2 == 1 : Should not set;

So far everything seems to working fine and it looks like h/w ignores these bits
for some reason. But this is not correct from programming perspective because
specification says (for example bank0) MSIIR address offset is 0x041740 but what
we were programming in device is 0x041747.

This also solves vfio-pci issue we were seeing on t4240.

Signed-off-by: Bharat Bhushan <Bharat.Bhushan@freescale.com>
Change-Id: Ibe4314f3149d6910608205b9d19f4dee334d8fc0
Reviewed-on: http://git.am.freescale.net:8181/6872
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Tiefei Zang <roy.zang@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>