dpaa_eth: Fix DMA unmapping of buffers on Rx
On the Rx path, we only DMA unmap the data buffer once we are
certain that we can allocate an associated skb. Otherwise we need
to keep the original mapping and release the buffer back into the
bpool.

This fixes a bug introduced by the patch that added ASF hooks in
the S/G driver code.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: I2fc5389727e94655113b52f31cad4b893b6a93f0
Reviewed-on: http://git.am.freescale.net:8181/6974
Reviewed-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Reviewed-by: Cristian-Constantin Sovaiala <Cristian.Sovaiala@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>