dpa_offload: Aggregate dynamically created HM operations
The DPA classifier driver allows multiple HM operations to be aggregated
in the same HM node only when importing the HM node resources. When
creating the nodes dynamically, the driver does not search for candidates
for aggregation before creating the HM nodes.

This update improves the dpa_classifier driver to actively check for
HM operation aggregation possibilities to certain neighbor HM nodes even
when using the API in dynamic mode.

Change-Id: Ic6cd0e280c2b6921718fd39047e22259f54c334b
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/15193
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>
Tested-by: Jose Rivera <German.Rivera@freescale.com>