dpa_offload: Fix update HM init failure when using IP frag only
Due to the recent aggregation that the dpa_classifier is performing for
chained header manip nodes, a problem has occured in the update HM
operation. This particular operation can consist of either one or two
HM nodes, depending on what HM options the user is selecting in the
parameters. This header manip operation can include an optional IP
fragmentation. It can also consist ONLY of IP fragmentation at user's
request. However, in this last case, the allocation of underlying HM
nodes was incorrect and it was fixed.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I753d1d6256f67b268c8c92f29aeb5d5f942cc311
Reviewed-on: http://git.am.freescale.net:8181/15622
Reviewed-by: Anca Jeanina Floarea <anca.floarea@freescale.com>
Reviewed-by: Richard Schmitt <richard.schmitt@freescale.com>
Tested-by: Richard Schmitt <richard.schmitt@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>