dpa_offload: Fix update HM init failure when using IP frag only
Due to the recent aggregation that the dpa_classifier is performing for
chained header manip nodes, a problem has occured in the update HM
operation. This particular operation can consist of either one or two
HM nodes, depending on what HM options the user is selecting in the
parameters. This header manip operation can include an optional IP
fragmentation. It can also consist ONLY of IP fragmentation at user's
request. However, in this last case, the allocation of underlying HM
nodes was incorrect and it was fixed.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: Idb1ebe3cb518e1fd7a9d52fb0423e9e0e9b7900c
Reviewed-on: http://git.am.freescale.net:8181/15621
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/19570
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Richard Schmitt <richard.schmitt@freescale.com>