dpa_offload: Avoid dereferencing an invalid pointer
In function "dpa_classif_mcast_create_group" from the dpa_classifier
driver there is a danger of dereferencing pgroup->entries pointer which
can be NULL in case of an error. To avoid this, in case of error this
pointer is checked before using it to release resources. Also, in
order to keep consistency of pgroup->entries in case of an allocation
error of pgroup->member_ids, the allocation of this structure and the
allocation of pgroup->member_ids have swicthed places.

Change-Id: Icfeb6cc3bb4b82949ec7e68741ac738924e191fd
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/19248
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>
Reviewed-by: Richard Schmitt <richard.schmitt@freescale.com>