dpa_offload: Remove dead code from "table_delete_entry_by_ref"
There is an unreachable code paragraph in function "table_delete_entry_by_ref".
The function doesn't work for indexed tables. In the first half, an error
is issued in this case. However, in the second half, another code paragraph
is testing again if the table type is "indexed". This branch will never
be reachable and was removed.

Change-Id: Ia5b565bc49130e4b0bc96a08ecdf37b6dcba44ed
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/19253
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Anca Jeanina Floarea <AncaJeanina.Floarea@freescale.com>
Reviewed-by: Richard Schmitt <richard.schmitt@freescale.com>