crypto: caamctrl: fix array out of bound access
desc_len(desc) returns a result masked by a seven bits mask representing
a descriptor length. The maximum length is 64 u32 words but masking
allows for larger but incorrect values.

This fix adds a bound check to the index of deco.descbuf array
(which is smaller than the maximum of 64 words), to avoid invalid
memory accesses.

Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Change-Id: If77757e64601eacba14f63f7c3782795b4993e2e
Reviewed-on: http://git.am.freescale.net:8181/20172
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Mircea Pop <mircea.pop@freescale.com>
Reviewed-by: Richard Schmitt <richard.schmitt@freescale.com>