Release overallocation of USDPAA memory
In certain versions of the Linux kernel memblock_alloc() silently rounds up
the allocation size. This can cause a significant loss in available memory
when large USDPAA DMA memory regions are configuired.

This patch caclulates the amount of the over allocation and returns the
unneeded memory back to the memblock allocator

Signed-off-by: Roy Pledge <Roy.Pledge@freescale.com>
Change-Id: I79ca00602fa04a7cd07caf9cf8f589864411fe45
Reviewed-on: http://git.am.freescale.net:8181/22454
Reviewed-by: Jeffrey Ladouceur <Jeffrey.Ladouceur@freescale.com>
Reviewed-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Reviewed-by: Richard Schmitt <richard.schmitt@freescale.com>
Tested-by: Richard Schmitt <richard.schmitt@freescale.com>