dpa_offload: Avoid creating SEC overlay commands on uninitialized memory
The sec_desc_extra_cmds memory space (which is part of the
"dpa_ipsec_sa" data structure) is allocated per IPSec tunnel and
aligned to L1 cache line once at the initialization of the SA manager.
This memory space is recycled for new tunnels without removing it
when the associated SA is removed and creating a new one with a new
SA. This process may leave garbage in this descriptor overlay
commands space and needs to be cleaned up with each new use.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>