fmd: Fix modification of nodes pointed by HM
There is a problem while executing modifications at runtime on
specific flows using header manipulations, especially in the presence
of traffic. When working with shadow mode for table management there
are two host commands that are executed. The second host command,
however was hiding an error in the first one which made it invisible
to tests unless the modifications were performed in the presence of
traffic.

The process of modification of nodes pointed by header manipulation
was updated to build the uCode configuration data structure correctly.

Signed-off-by: Ella Segal <ella.segal@freescale.com>
Tested-by: Marian Chereji <marian.chereji@freescale.com>