vfio fsl-mc: Temporary: Extend VFIO-API to pass guest irq-num
This patch allows userspace/qemu to supply an extra
parameter "irq_num", which VFIO driver writes to mc-object.

Later when guest boots then it reads the "irq-num" from MC and
uses this as h/w irq-number for setting up interrupt for the object.
This is needed for supporting Line interrupt in guest and will
be removed when dprc will be emulated in QEMU.

Signed-off-by: Bharat Bhushan <Bharat.Bhushan@freescale.com>