KVM arm: Increase the Memslots and expose to usespace
We are direct assigning the the fsl-mc container which have
many devices and these devices have memory regions.
We are also working on direct assigning these fsl-mc containers
to userspace in VM, example enabling ODP/DPDK in guest VM.

So overall the number of device memory regions are more and
32 memslots are not sufficient. So this patch increases the
number of memslots from 32 to 54.

Also this patch add support of exposing these memslots count
to user-sapce.

Signed-off-by: Bharat Bhushan <Bharat.Bhushan@freescale.com>